+++ src/de/hhu/bsinfo/dxcompute/job/AbstractJobComponent.java
@@ -16,7 +16,7 @@ public abstract class AbstractJobComponent extends AbstractDXRAMComponent {
 	 * @param p_priorityInit     Default init priority for this component
 	 * @param p_priorityShutdown Default shutdown priority for this component
 	 */
-	public AbstractJobComponent(final int p_priorityInit, final int p_priorityShutdown) {
+	public AbstractJobComponent(final short p_priorityInit, final short p_priorityShutdown) {
 		super(p_priorityInit, p_priorityShutdown);
 	}
